requirements.txt
streamlit>=1.24.0
plotly>=5.13.0
pandas>=1.5.0
numpy
python-dotenv>=0.21.0
supabase>=1.0.3
requests>=2.28.0
httpx>=0.24.0
